Написал User в Application

+++ IdentityLearning.Identity/Repositories/UserRepository/AdminUserRepository.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using IdentityLearning.Application.Contracts.Identity.UserRepository;
 using IdentityLearning.Application.Exceptions;
 using IdentityLearning.Domain.Entities.User;
+using IdentityLearning.Domain.Models;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 
@@ -16,20 +17,28 @@ namespace IdentityLearning.Identity.Repositories.UserRepository
         private readonly UserManager<ApplicationUser> _userManager;
         private readonly IdentityDbContext _dbContext;
 
-        public AdminUserRepository(UserManager<ApplicationUser> userManager, IdentityDbContext dbContext) : base(userManager)
+        public AdminUserRepository(UserManager<ApplicationUser> userManager, IdentityDbContext dbContext) : base(userManager, dbContext)
         {
             _userManager = userManager;
             _dbContext = dbContext;
         }
 
 
-        public async Task DeleteUser(ApplicationUser user)
+        public async Task<Result<object>> DeleteUser(ApplicationUser user)
         {
+            var isAdmin = await _userManager.IsInRoleAsync(user, "Admin");
+            if (isAdmin)
+            {
+                return Result<object>.NotSuccessfull("Can't delete admin user", TotalErrorCode.Forbidden);
+            }
+
             var deleteResult =  await _userManager.DeleteAsync(user);
             if (deleteResult.Succeeded == false)
             {
                 throw new IdentityException(deleteResult);
             }
+
+            return Result<object>.Successfull(null!);
         }
 
         public async Task DeleteAllUsers()
